Bug fix and use CreateFileW Fixed bug where I CloseHandle before GetLastError() causing INVALID_HANDLE_VALUE as the HANDLE wasn't likely created in the first place. Changed CreateFile to CreateFileW although CreateFile expands to CreateFileW I want to explicitly specify CreateFileW just incase

+++ touch/touch.h
@@ -63,7 +63,7 @@ public:
     std::wstring FileReference() { return m_Settings.FileReference; };
     void FileReference(std::wstring FileReference) { m_Settings.FileReference = FileReference; };
     void AddFile(std::wstring FileName) { m_Settings.FileNames.push_back(FileName); }
-    int32_t FileCount() { return m_Settings.FileNames.size(); }
+    size_t FileCount() { return m_Settings.FileNames.size(); }
     bool GetDateFromString(std::wstring& stamp, SYSTEMTIME& time);
     bool GetTimeFromString(std::wstring& stamp, SYSTEMTIME& time);
     int32_t touch();
